pyproject.toml
[tool.poetry]
name = "django project"
version = "0.1.0"
description = ""
authors = ["natalia_levant <an_smir@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
asgiref = "^3.8.1"
sqlparse = "^0.5.3"
pillow = "^11.0.0"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
docutils = "^0.21.2"
flake8 = "^7.1.1"
flake8-docstrings = "^1.7.0"
black = "^24.10.0"
drf-spectacular = "^0.28.0"
django-debug-toolbar = "^4.4.6"
gunicorn = "^23.0.0"
isort = "^5.13.2"

[tool.black]
line-length = 90
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 90
multi_line_output = 3
skip_gitignore = true
skip_glob = ["**/migrations/*", "**/settings/*"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
